Refactor path
